Move TransformationIdentity to right project and package
